requirements.txt
click
Pillow
numpy
scipy
